add more newline tests
